+++ uniel_protocol.h
@@ -13,10 +13,9 @@ public:
     static const int DefaultTimeoutMs = 1000;
 
     TUnielProtocol(PAbstractSerialPort port);
-    uint64_t ReadRegister(uint8_t mod, uint8_t address, RegisterFormat fmt);
-    void WriteRegister(uint8_t mod, uint8_t address, uint64_t value, RegisterFormat fmt);
-
-    void SetBrightness(uint8_t mod, uint8_t address, uint8_t value);
+    uint64_t ReadRegister(uint32_t mod, uint32_t address, RegisterFormat fmt);
+    void WriteRegister(uint32_t mod, uint32_t address, uint64_t value, RegisterFormat fmt);
+    void SetBrightness(uint32_t mod, uint32_t address, uint8_t value);
 
 private:
     void WriteCommand(uint8_t cmd, uint8_t mod, uint8_t b1, uint8_t b2, uint8_t b3);
